Controller-10.30.7.52/ip-sdn-controller/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
kafka-python==2.0.2
networkx==3.1
python-dotenv==1.0.0
